requirements.txt
streamlit
pandas
numpy
plotly
scipy
anthropic
python-dotenv
